extract styledmenu as seperate component
