+++ frontend/src/components/header/LanguageSelect.js
@@ -1,4 +1,4 @@
-import { Button, makeStyles, Menu, MenuItem, useMediaQuery, withStyles } from "@material-ui/core";
+import { Button, makeStyles, useMediaQuery } from "@material-ui/core";
 import LanguageIcon from "@material-ui/icons/Language";
 import { useRouter } from "next/router";
 import React, { useContext, useEffect, useRef, useState } from "react";
@@ -6,6 +6,8 @@ import Cookies from "universal-cookie";
 import { getCookieProps } from "../../../public/lib/cookieOperations";
 import theme from "../../themes/theme";
 import UserContext from "../context/UserContext";
+import StyledMenu from "../general/StyledMenu";
+import StyledMenuItem from "../general/StyledMenuItem";
 
 const useStyles = makeStyles((theme) => ({
   root: (props) => ({
@@ -115,34 +117,3 @@ export default function LanguageSelect({ transparentHeader }) {
     </>
   );
 }
-
-const StyledMenu = withStyles({
-  paper: {
-    width: 64,
-  },
-})((props) => (
-  <Menu
-    elevation={0}
-    getContentAnchorEl={null}
-    anchorOrigin={{
-      vertical: "bottom",
-      horizontal: "center",
-    }}
-    transformOrigin={{
-      vertical: "top",
-      horizontal: "center",
-    }}
-    {...props}
-  />
-));
-
-const StyledMenuItem = withStyles(() => ({
-  root: {
-    color: "primary",
-    textAlign: "center",
-    fontWeight: 600,
-  },
-  selected: {
-    color: "white",
-  },
-}))(MenuItem);
